# api/openapi/openapi.yaml
openapi: 3.0.4
info:
  title: 'Meals API'
  description: 'API for family meals management'
  contact:
    email: 'benditdev90@gmail.com'
    name: 'Owner'
  version: '1.0.0'
servers:
  - url: 'https://fooder.bendit.dev/api/v1'
tags:
  - name: 'ingredient'
  - name: 'recipe'
  - name: 'cooking-list'
paths:
  /ingredients:
    get:
      description: 'Returns a collection of ingredients'
      operationId: 'getIngredients'
      tags: ['ingredient']
      responses:
        '200':
          description: 'A collection of ingredients'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      description: 'Creates a new ingredient'
      operationId: 'createIngredient'
      tags: ['ingredient']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientCreate'
      responses:
        '201':
          description: 'Ingredient created'
          headers:
            Location:
              description: location of newly created resource
              schema:
                type: string
                example: '/ingredients/1'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /ingredients/{id}:
    get:
      description: 'Returns an ingredient resource'
      operationId: 'getIngredientByID'
      tags: ['ingredient']
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'An ingredient resource'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /recipes:
    get:
      description: 'Returns a collection of recipes'
      operationId: 'getRecipes'
      tags: ['recipe']
      responses:
        '200':
          description: 'A collection of recipes'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      description: 'Creates new recipe'
      operationId: 'createRecipe'
      tags: ['recipe']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
      responses:
        '201':
          description: 'Recipe created'
          headers:
            Location:
              description: location of newly created resource
              schema:
                type: string
                example: '/recipes/1'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /recipes/{id}:
    get:
      description: 'Returns a recipe resource'
      operationId: 'getRecipeByID'
      tags: ['recipe']
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'A recipe resource'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cooking-lists:
    post:
      description: 'Creates a new cooking list with an initial recipe'
      operationId: 'createCookingList'
      tags: ['cooking-list']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCookingListRequest'
      responses:
        '201':
          description: 'Cooking list created'
          headers:
            Location:
              description: location of newly created resource
              schema:
                type: string
                example: '/cooking-lists/1'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cooking-lists/{id}:
    get:
      description: 'Returns a cooking list resource'
      operationId: 'getCookingListByID'
      tags: ['cooking-list']
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'A cooking list resource'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookingListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      description: 'Adds a recipe to an existing cooking list'
      operationId: 'addRecipeToCookingList'
      tags: ['cooking-list']
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRecipeToCookingListRequest'
      responses:
        '202':
          description: 'Recipe added to cooking list'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NotAuthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 'Could not add recipe to cooking list'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemJsonDetails'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /cooking-lists/{id}/shopping-list:
    get:
      description: 'Generates a shopping list from a cooking list'
      operationId: 'generateShoppingList'
      tags: ['cooking-list']
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'A shopping list generated from the cooking list'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShoppingListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  responses:
    BadRequest:
      description: 'Invalid request payload or query params'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemJsonDetails'

    NotAuthenticated:
      description: 'User did not authenticate, login or valid api key required'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemJsonDetails'

    Forbidden:
      description: 'User is not authorized to perfom the operation'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemJsonDetails'

    NotFound:
      description: 'Requested resource was not found'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemJsonDetails'

    InternalServerError:
      description: 'Unexpected server error occurred'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemJsonDetails'

  schemas:
    Unit:
      type: string
      enum: [g, kg, ml, l, piece, tbsp, tsp]
    Ingredient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    IngredientCreate:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    RecipeIngredient:
      type: object
      properties:
        ingredientId:
          type: integer
        name:
          type: string
        amount:
          type: number
          format: double
        unit:
          $ref: '#/components/schemas/Unit'
      required:
        - ingredientId
        - name
        - amount
        - unit
    RecipeIngredientCreate:
      type: object
      properties:
        ingredientId:
          type: integer
        amount:
          type: number
          format: double
        unit:
          $ref: '#/components/schemas/Unit'
      required:
        - ingredientId
        - amount
        - unit
    Recipe:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredient'
      required:
        - id
        - title
        - description
        - ingredients
    RecipeCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/RecipeIngredientCreate'
      required:
        - title
        - description
        - ingredients
    CreateCookingListRequest:
      type: object
      properties:
        recipeId:
          type: integer
      required:
        - recipeId
    AddRecipeToCookingListRequest:
      type: object
      properties:
        recipeId:
          type: integer
      required:
        - recipeId
    CookingListItem:
      type: object
      properties:
        recipeId:
          type: integer
    CookingListResponse:
      type: object
      properties:
        id:
          type: integer
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/CookingListItem'
    ShoppingListItem:
      type: object
      properties:
        ingredientId:
          type: integer
        name:
          type: string
        totalAmount:
          type: number
          format: double
        unit:
          $ref: '#/components/schemas/Unit'
      required:
        - ingredientId
        - name
        - totalAmount
        - unit
    ShoppingListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShoppingListItem'
      required:
        - items
    ProblemJsonDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri